+++ src/components/PeopleTable.tsx
@@ -3,8 +3,9 @@ import classNames from 'classnames';
 import { useMemo } from 'react';
 import { PersonLink } from './Links';
 import { Person } from '../types';
-import { Params, getSearchWith } from '../utils/getSearchWith';
+import { getSearchWith } from '../utils/getSearchWith';
 import { SortType } from '../types/SortType';
+import { getFiltered } from '../utils/getFiltered';
 
 type Props = {
   people: Person[],
@@ -14,72 +15,6 @@ type Props = {
   sexFilter: string,
   centuries: string[],
 };
-
-function getFiltered(
-  people: Person[],
-  {
-    query,
-    sexFilter,
-    centuries,
-    sort,
-    order,
-  }: Params,
-) {
-  let filteredPeople = [...people];
-
-  if (sexFilter) {
-    filteredPeople = filteredPeople.filter(person => {
-      return person.sex === sexFilter;
-    });
-  }
-
-  if (query && typeof query === 'string') {
-    const correctQuery = query.trim().toLowerCase();
-
-    filteredPeople = filteredPeople.filter(person => {
-      return (
-        person.name.toLowerCase().includes(correctQuery)
-        || person.motherName?.toLowerCase().includes(correctQuery)
-        || person.fatherName?.toLowerCase().includes(correctQuery)
-      );
-    });
-  }
-
-  if (centuries
-    && Array.isArray(centuries)
-    && centuries.length !== 0) {
-    const correctCentury = centuries.map(century => +century);
-
-    filteredPeople = filteredPeople.filter(person => {
-      const bornCentury = Math.floor(person.born / 100) + 1;
-
-      return correctCentury.includes(bornCentury);
-    });
-  }
-
-  if (sort) {
-    filteredPeople = filteredPeople.sort((person1, person2) => {
-      switch (sort) {
-        case SortType.NAME:
-        case SortType.SEX:
-          return person1[sort].localeCompare(person2[sort]);
-
-        case SortType.BORN:
-        case SortType.DIED:
-          return person1[sort] - person2[sort];
-
-        default:
-          return 0;
-      }
-    });
-  }
-
-  if (order) {
-    filteredPeople = filteredPeople.reverse();
-  }
-
-  return filteredPeople;
-}
 
 export const PeopleTable: React.FC<Props> = ({
   people,
