+++ src/components/constants/FilterNumbers.ts
@@ -1,0 +1,1 @@
+export const centuryFilterNumbers = ['16', '17', '18', '19', '20'];
